move connection options
